--- pkg/client/datadog_wrapper.go
@@ -1,214 +1,0 @@
-package client
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-
-	"github.com/DataDog/datadog-api-client-go/v2/api/datadog"
-	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV1"
-	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV2"
-)
-
-// DatadogWrapper wraps the Datadog REST client to provide a clean interface for all endpoints.
-// This wrapper delegates to the DatadogRestClient which already handles HTTP response body closing correctly.
-type DatadogWrapper struct {
-	restClient *DatadogRestClient
-	// Keep reference to official client for operations that can't be implemented in REST client yet
-	officialClient *datadog.APIClient
-}
-
-// NewDatadogWrapper creates a new wrapper that uses the custom REST client.
-func NewDatadogWrapper(restClient *DatadogRestClient, officialClient *datadog.APIClient) *DatadogWrapper {
-	return &DatadogWrapper{
-		restClient:     restClient,
-		officialClient: officialClient,
-	}
-}
-
-// GetRestClient returns the underlying REST client.
-func (w *DatadogWrapper) GetRestClient() *DatadogRestClient {
-	return w.restClient
-}
-
-// GetOfficialClient returns the underlying official Datadog API client for operations that need it.
-func (w *DatadogWrapper) GetOfficialClient() *datadog.APIClient {
-	return w.officialClient
-}
-
-// ListUsers lists users using the REST client.
-func (w *DatadogWrapper) ListUsers(ctx context.Context, params *datadogV2.ListUsersOptionalParameters) (*datadogV2.UsersResponse, error) {
-	usersApi := datadogV2.NewUsersApi(w.officialClient)
-	var (
-		resp    datadogV2.UsersResponse
-		httpRes *http.Response
-		err     error
-	)
-	if params != nil {
-		resp, httpRes, err = usersApi.ListUsers(ctx, *params)
-	} else {
-		resp, httpRes, err = usersApi.ListUsers(ctx)
-	}
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	if err != nil {
-		return nil, err
-	}
-	return &resp, nil
-}
-
-// Validate validates API credentials using the REST client.
-func (w *DatadogWrapper) Validate(ctx context.Context) (*datadogV1.AuthenticationValidationResponse, error) {
-	// TODO: Implement in DatadogRestClient
-	// For now, return error indicating this needs to be implemented
-	return nil, fmt.Errorf("Validate not yet implemented in REST client")
-}
-
-// ListTeams lists teams using the REST client.
-func (w *DatadogWrapper) ListTeams(ctx context.Context, params *datadogV2.ListTeamsOptionalParameters) (*datadogV2.TeamsResponse, error) {
-	teamsApi := datadogV2.NewTeamsApi(w.officialClient)
-	var (
-		resp    datadogV2.TeamsResponse
-		httpRes *http.Response
-		err     error
-	)
-	if params != nil {
-		resp, httpRes, err = teamsApi.ListTeams(ctx, *params)
-	} else {
-		resp, httpRes, err = teamsApi.ListTeams(ctx)
-	}
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	if err != nil {
-		return nil, err
-	}
-	return &resp, nil
-}
-
-// ListRoles lists roles using the REST client.
-func (w *DatadogWrapper) ListRoles(ctx context.Context, params *datadogV2.ListRolesOptionalParameters) (*datadogV2.RolesResponse, error) {
-	rolesApi := datadogV2.NewRolesApi(w.officialClient)
-	var (
-		resp    datadogV2.RolesResponse
-		httpRes *http.Response
-		err     error
-	)
-	if params != nil {
-		resp, httpRes, err = rolesApi.ListRoles(ctx, *params)
-	} else {
-		resp, httpRes, err = rolesApi.ListRoles(ctx)
-	}
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	if err != nil {
-		return nil, err
-	}
-	return &resp, nil
-}
-
-// ListAPIKeys lists API keys using the REST client.
-func (w *DatadogWrapper) ListAPIKeys(ctx context.Context, params *datadogV2.ListAPIKeysOptionalParameters) (*datadogV2.APIKeysResponse, error) {
-	keysApi := datadogV2.NewKeyManagementApi(w.officialClient)
-	var (
-		resp    datadogV2.APIKeysResponse
-		httpRes *http.Response
-		err     error
-	)
-	if params != nil {
-		resp, httpRes, err = keysApi.ListAPIKeys(ctx, *params)
-	} else {
-		resp, httpRes, err = keysApi.ListAPIKeys(ctx)
-	}
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	if err != nil {
-		return nil, err
-	}
-	return &resp, nil
-}
-
-// Wrapper methods that handle HTTP response body closing automatically
-
-// ListRoleUsers lists users for a specific role and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) ListRoleUsers(ctx context.Context, roleId string, params *datadogV2.ListRoleUsersOptionalParameters) (*datadogV2.UsersResponse, error) {
-	rolesApi := datadogV2.NewRolesApi(w.officialClient)
-	users, httpRes, err := rolesApi.ListRoleUsers(ctx, roleId, *params)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &users, err
-}
-
-// AddUserToRole adds a user to a role and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) AddUserToRole(ctx context.Context, roleId string, body datadogV2.RelationshipToUser) (*datadogV2.UsersResponse, error) {
-	rolesApi := datadogV2.NewRolesApi(w.officialClient)
-	resp, httpRes, err := rolesApi.AddUserToRole(ctx, roleId, body)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &resp, err
-}
-
-// RemoveUserFromRole removes a user from a role and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) RemoveUserFromRole(ctx context.Context, roleId string, body datadogV2.RelationshipToUser) (*datadogV2.UsersResponse, error) {
-	rolesApi := datadogV2.NewRolesApi(w.officialClient)
-	resp, httpRes, err := rolesApi.RemoveUserFromRole(ctx, roleId, body)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &resp, err
-}
-
-// GetTeamMemberships gets team memberships and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) GetTeamMemberships(ctx context.Context, teamId string, params *datadogV2.GetTeamMembershipsOptionalParameters) (*datadogV2.UserTeamsResponse, error) {
-	teamsApi := datadogV2.NewTeamsApi(w.officialClient)
-	memberships, httpRes, err := teamsApi.GetTeamMemberships(ctx, teamId, *params)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &memberships, err
-}
-
-// GetUser gets a user by ID and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) GetUser(ctx context.Context, userId string) (*datadogV2.UserResponse, error) {
-	usersApi := datadogV2.NewUsersApi(w.officialClient)
-	user, httpRes, err := usersApi.GetUser(ctx, userId)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &user, err
-}
-
-// CreateTeamMembership creates a team membership and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) CreateTeamMembership(ctx context.Context, teamId string, body datadogV2.UserTeamRequest) (*datadogV2.UserTeamResponse, error) {
-	teamsApi := datadogV2.NewTeamsApi(w.officialClient)
-	resp, httpRes, err := teamsApi.CreateTeamMembership(ctx, teamId, body)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &resp, err
-}
-
-// DeleteTeamMembership deletes a team membership and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) DeleteTeamMembership(ctx context.Context, teamId string, userId string) error {
-	teamsApi := datadogV2.NewTeamsApi(w.officialClient)
-	httpRes, err := teamsApi.DeleteTeamMembership(ctx, teamId, userId)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return err
-}
-
-// ValidateCredentials validates API credentials and automatically handles HTTP response body closing.
-func (w *DatadogWrapper) ValidateCredentials(ctx context.Context) (*datadogV1.AuthenticationValidationResponse, error) {
-	api := datadogV1.NewAuthenticationApi(w.officialClient)
-	resp, httpRes, err := api.Validate(ctx)
-	if httpRes != nil {
-		defer httpRes.Body.Close()
-	}
-	return &resp, err
-}
